pyproject.toml
[tool.poetry]
name = "bank-project"
version = "0.1.0"
description = "Banking API"
authors = ["Vikrom Narula <vikrom.nar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
faker = "^22.0.0"
autopep8 = "^2.0.4"
django-stubs = "^4.2.7"
django = "^5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120
ignore = ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3
